CIV-13628 update api tests
