+++ e2e/tests/api_tests/sdo_R2/sdo_flight_delay_test.js
@@ -2,23 +2,28 @@ const config = require('../../../config.js');
 // To use on local because the idam images are different
 // const judgeUser = config.judgeUserWithRegionId1Local;
 // const hearingCenterAdminToBeUsed = config.hearingCenterAdminLocal;
-const judgeUser = config.testEarlyAdopterCourts ? config.judgeUser2WithRegionId2 : config.judgeUserWithRegionId1;
-const hearingCenterAdminToBeUsed = config.testEarlyAdopterCourts ? config.hearingCenterAdminWithRegionId2 : config.hearingCenterAdminWithRegionId1;
-
-
+const judgeUser = config.judgeUser2WithRegionId4; //small claim specified goes to region4 judge
+const hearingCenterAdminToBeUsed = config.hearingCenterAdminWithRegionId4;
+const mpScenario = 'ONE_V_ONE_FLIGHT_DELAY';
 const mpScenario1v1 = 'ONE_V_ONE';
+const claimAmountSmallTrack = '1500';
 
-async function prepareClaimSpec(api_spec) {
-  await api_spec.createClaimWithRepresentedRespondent(config.applicantSolicitorUser, 'ONE_V_ONE');
-  await api_spec.defendantResponse(config.defendantSolicitorUser, 'FULL_DEFENCE', 'ONE_V_ONE', true);
-  await api_spec.claimantResponse(config.applicantSolicitorUser, 'FULL_DEFENCE', 'ONE_V_ONE',
-    'AWAITING_APPLICANT_INTENTION');
+async function prepareClaim(api_spec, claimAmount) {
+  await api_spec.createClaimSpecFlightDelay(config.applicantSolicitorUser, mpScenario, claimAmount);
+  await api_spec.defendantResponse(config.defendantSolicitorUser);
+  await api_spec.claimantResponseForFlightDelay(config.applicantSolicitorUser, 'FULL_DEFENCE', 'ONE_V_ONE', 'AWAITING_APPLICANT_INTENTION');
+}
+
+async function prepareClaimOtherOption(api_spec, claimAmount) {
+  await api_spec.createClaimSpecFlightDelay(config.applicantSolicitorUser, 'ONE_V_ONE_FLIGHT_DELAY_OTHER', claimAmount);
+  await api_spec.defendantResponse(config.defendantSolicitorUser);
+  await api_spec.claimantResponseForFlightDelay(config.applicantSolicitorUser, 'FULL_DEFENCE', 'ONE_V_ONE', 'AWAITING_APPLICANT_INTENTION');
 }
 
 Feature('Create SDO SmallTrack- Flight Delay - 1v1 - spec @api-specified @api-nonprod');
 Scenario('1v1 full defence unspecified - judge draws small claims WITHOUT sum of damages - flight delay', async ({api_spec}) => {
   if (['preview', 'demo'].includes(config.runningEnv)) {
-    await prepareClaimSpec(api_spec);
+    await prepareClaim(api_spec, claimAmountSmallTrack);
     await api_spec.createSDO(judgeUser, 'CREATE_SMALL');
     await api_spec.evidenceUploadApplicant(config.applicantSolicitorUser, mpScenario1v1);
     await api_spec.evidenceUploadRespondent(config.defendantSolicitorUser, mpScenario1v1);
@@ -32,6 +37,13 @@ Scenario('1v1 full defence unspecified - judge draws small claims WITHOUT sum of
   }
 });
 
-AfterSuite(async ({api_spec_small}) => {
-  await api_spec_small.cleanUp();
+Scenario('1v1 specified - flight delay other option Small-claim', async ({api_spec}) => {
+  if (['preview', 'demo'].includes(config.runningEnv)) {
+    await prepareClaimOtherOption(api_spec, claimAmountSmallTrack);
+  }
+});
+
+
+AfterSuite(async ({api_spec}) => {
+  await api_spec.cleanUp();
 });
